src/test/java/org/sourav/project/personal/Controller/ControllerTest.java
package org.sourav.project.personal.Controller;

import org.hamcrest.Matchers;
import org.junit.Before;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.MockitoAnnotations;
import org.sourav.project.personal.Service.Distance;
import org.sourav.project.personal.bean.ApiResponse;
import org.sourav.project.personal.bean.Request;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.mock.web.MockHttpServletResponse;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.web.context.WebApplicationContext;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
import static org.springframework.test.web.servlet.setup.MockMvcBuilders.webAppContextSetup;

@ExtendWith(SpringExtension.class)
@WebMvcTest(Controller.class)
class ControllerTest {

    @Autowired
    private WebApplicationContext webApplicationContext;

    @Autowired
    private MockMvc mockMvc;


    @MockBean
    private Distance mockDistance;

    @Before
    public void init(){
        MockitoAnnotations.initMocks(this);
    }



    @Test
    void testFindDistance_DistanceThrowsException() throws Exception {
        // Setup
        when(mockDistance.getDistance(any(Request.class))).thenThrow(Exception.class);

        // Run the test
        final MockHttpServletResponse response = mockMvc.perform(get("/findDistance")
                        .header("content-type", "contentType")
                        .accept(MediaType.APPLICATION_JSON))
                .andReturn().getResponse();

        // Verify the results
        assertEquals(HttpStatus.INTERNAL_SERVER_ERROR.value(), response.getStatus());
        //assertEquals("expectedResponse", response.getContentAsString());
    }
}
